requirements.txt
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
Flask==3.1.2
Flask-Session==0.8.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
msgspec==0.20.0
openai==2.16.0
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
